+++ CHANGELOG.md
@@ -1,7 +1,13 @@
-## metaheuristic_algorithms_python 0.1.0 (October 26, 2015) ##
+## metaheuristic_algorithms_python 0.1.2 (November 9, 2015) ##
 
-* Added Harmony Search
+* Fixed a typo in method name minimum_decision_variable_values in FunctionWrapper's.
+
+* Added Simulated Annealing
 
 ## metaheuristic_algorithms_python 0.1.1 (November 3, 2015) ##
 
 * Added Simplified Particle Swarm Optimization
+
+## metaheuristic_algorithms_python 0.1.0 (October 26, 2015) ##
+
+* Added Harmony Search